--- a/bench_helpers/Cargo.toml
+++ b/bench_helpers/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "bench_helpers"
+version = "0.1.0"
+authors = ["dmitry.dygalo <dadygalo@gmail.com>"]
+edition = "2018"
+license = "MIT"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+criterion = "0.3"