As mentioned in #247 this PR splits the current benchmarks to avoid too long recompilation.

**Before** (`cargo bench --bench jsonschema`)
First compilation: 1m 36s
Re-compilation: 36s

**After** (`cargo bench --bench jsonschema`)
First compilation: 1m 32s
Re-compilation: 3.56s

Though, compilation of `jsonschema` and `jsonschema(bin)` are still slow.

TODO:
 - [x] Create common code to run benchmarks (loading schemas & their instances)